// ECommerceApp/ECommerceApp/src/common/types/types.ts
export enum UserRoles {
  user = 'user',
  admin = 'admin',
}

export enum UserGender {
  male = 'male',
  female = 'female',
}

export enum PaymentMethodTypes {
  cash = 'cash',
  card = 'card',
}

export enum OrderStatusTypes {
  pending = 'pending',
  confirmed = 'confirmed',
  shipped = 'shipped',
  delivered = 'delivered',
}